TFloat = Single (single precision everywhere) -- QuantizePalette: do not sort palettes colors by luma (Q boost)

+++ extern.pas
@@ -8,7 +8,7 @@ uses
   LazLogger, Windows, Classes, SysUtils, Types, Process, strutils, math;
 
 type
-  TFloat = Double;
+  TFloat = Single;
 
   TFloatDynArray = array of TFloat;
   TFloatDynArray2 = array of TFloatDynArray;
